Unpin Shared Notes before sharing External Video
